Move tests to tests/
