+++ tests/player_test.py
@@ -1,4 +1,4 @@
-from py_loto.player import Player
+from .py_loto.player import Player
 
 
 def test_instantiating():
