+++ tests/printer_test.test_print_player_with_no_cards.snapshot
