+++ tests/printer_test.test_print_player_with_partially_filled_card.snapshot
