+++ tests/printer_test.test_print_player_with_single_card.snapshot
